:sparkles: Removed scrollIntoView logic out of the work page render component into it's own component. It's now reusable, only needs targetID and description
